fix(ci): python-full-coverage - split oversized shards and cover parser branches
